change alwaysBalancing and always5vRequest to uint8_t

+++ Inc/main.h
@@ -146,8 +146,8 @@ typedef struct _ADCparameters {
 
 typedef struct _generalParameters {
 	uint16_t stayActiveTime;	//h (Hours), how long to stay in active mode
-	uint16_t alwaysBalancing;	//1 or 0, allow balancing even when not charging
-	uint16_t always5vRequest;	//1 or 0, force 5V buck always on
+	uint8_t alwaysBalancing;	//1 or 0, allow balancing even when not charging
+	uint8_t always5vRequest;	//1 or 0, force 5V buck always on
 
 	uint16_t storageCellVoltage;	//mV (milliVolts), what voltage to discharge cells if storage voltage discharge enabled
 	uint16_t timeToStorageDischarge;	//h (hours), how long to wait after active state to start discharging cells to storage voltage
